moved lattice embedder to embedders module
